sensor_checker/requirements.txt
confluent_kafka
requests
python-dateutil
pymongo
pydantic